#35: Rename move layer data into a new package
