+++ tf-layer-python/src/main/kotlin/edu/wpi/axon/tflayer/python/DefaultLayerToPythonCode.kt
@@ -1,8 +1,8 @@
 package edu.wpi.axon.tflayer.python
 
-import edu.wpi.axon.tflayers.Activation
-import edu.wpi.axon.tflayers.Layer
-import edu.wpi.axon.tflayers.SealedLayer
+import edu.wpi.axon.tfdata.layer.Activation
+import edu.wpi.axon.tfdata.layer.Layer
+import edu.wpi.axon.tfdata.layer.SealedLayer
 
 /**
  * An implementation of [LayerToCode] for Python.
