+++ tf-layer-python/src/main/kotlin/edu/wpi/axon/tflayer/python/LayerToCode.kt
@@ -1,7 +1,7 @@
 package edu.wpi.axon.tflayer.python
 
-import edu.wpi.axon.tflayers.Activation
-import edu.wpi.axon.tflayers.Layer
+import edu.wpi.axon.tfdata.layer.Activation
+import edu.wpi.axon.tfdata.layer.Layer
 
 interface LayerToCode {
